Extract CategoryGraph component
