+++ src/components/BudgetOverview.svelte
@@ -1,30 +1,7 @@
 <script>
+import CategoryGraph from './CategoryGraph.svelte'
 import { categories } from '../data/categories'
-import { formatAmount, formatAmountAsWholeNumber } from '../helpers/number-formats'
-
-const dangerIfNegative = remaining => isNegative(remaining) ? 'danger' : ''
-
-const isNegative = number => (number && (number < 0))
-
-const getStatus = (remaining, total) => {
-  if (remaining == undefined) {
-    return '';
-  } else if (remaining < 0) {
-    return 'danger';
-  } else if (remaining < (total / 4)) {
-    return 'warning';
-  }
-  return 'success';
-}
-
-const calculageWidth = (remaining, total) => {
-  if (remaining < 0) {
-    return 0;
-  } else if (total === 0) {
-    return (remaining > 0) ? 100 : 0;
-  }
-  return (remaining / total) * 100;
-}
+import { dangerIfNegative, formatAmount, formatAmountAsWholeNumber } from '../helpers/numbers'
 </script>
 
 <style>
@@ -47,33 +24,6 @@ const calculageWidth = (remaining, total) => {
 	vertical-align: bottom;
 }
 
-.category-graph {
-  box-shadow: 0px 0px 1px 0px rgba(0, 0, 0, 0.75);
-}
-.category-graph.danger {
-  box-shadow: 0px 0px 0px 1px rgba(255, 0, 0, 0.75);
-}
-
-.category-graph,
-.category-graph-line {
-  border-radius: 4px;
-}
-
-.category-graph .danger {
-  background-color: red;
-}
-.category-graph .success {
-  background-color: green;
-}
-.category-graph .warning {
-  background-color: orange;
-}
-
-.category-graph-line {
-	height: 6px;
-	margin: auto 0;
-}
-
 .category-list td {
   border: none;
   vertical-align: middle;
@@ -81,11 +31,6 @@ const calculageWidth = (remaining, total) => {
 
 .category-name {
   white-space: nowrap;
-}
-
-.category-name > * {
-	max-width: 100%;
-	overflow: hidden;
 }
 
 .width-10 {
@@ -106,10 +51,7 @@ const calculageWidth = (remaining, total) => {
              class="btn btn-outline-secondary">{ category.name }</a>
         </td>
         <td class="width-80">
-          <div class="category-graph { dangerIfNegative(category.remaining) }">
-            <div class="category-graph-line { getStatus(category.remaining, category.amount) }"
-                 style="width: { calculageWidth(category.remaining, category.budgetedAmount) }%;"></div>
-          </div>
+          <CategoryGraph {category} />
         </td>
         <td class="category-amount width-10">
           <div class="category-available { dangerIfNegative(category.remaining) }">
