+++ src/helpers/numbers.js
@@ -1,4 +1,6 @@
 
+export const dangerIfNegative = remaining => isNegative(remaining) ? 'danger' : ''
+
 export const formatAmount = amount => formatAmountWithPrecision(amount, 2)
 
 export const formatAmountAsWholeNumber = amount => formatAmountWithPrecision(amount, 0)
@@ -9,3 +11,5 @@ const formatAmountWithPrecision = (amount, precision) => {
   }
   return (Number(amount) / 100).toFixed(precision)
 }
+
+const isNegative = number => (number && (number < 0))
